Merge hierarchy store into Tachyon [TACHYON-33 sub-task 5 & 6]

+++ core/src/main/java/tachyon/client/BlockOutStream.java
@@ -27,6 +27,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import tachyon.Constants;
+import tachyon.thrift.ClientLocationInfo;
 import tachyon.util.CommonUtils;
 
 /**
@@ -43,6 +44,7 @@ public class BlockOutStream extends OutStream {
 
   private long mInFileBytes = 0;
   private long mWrittenBytes = 0;
+  private ClientLocationInfo mLocationInfo = null;
 
   private String mLocalFilePath = null;
   private RandomAccessFile mLocalFile = null;
@@ -81,28 +83,19 @@ public class BlockOutStream extends OutStream {
       throw new IOException(msg);
     }
 
-    File localFolder = mTachyonFS.createAndGetUserLocalTempFolder();
-    if (localFolder == null) {
-      mCanWrite = false;
-      String msg = "Failed to create temp user folder for tachyon client.";
-      throw new IOException(msg);
-    }
-
-    mLocalFilePath = CommonUtils.concat(localFolder.getPath(), mBlockId);
-    mLocalFile = new RandomAccessFile(mLocalFilePath, "rw");
-    mLocalFileChannel = mLocalFile.getChannel();
-    // change the permission of the temporary file in order that the worker can move it.
-    CommonUtils.changeLocalFileToFullPermission(mLocalFilePath);
-    // use the sticky bit, only the client and the worker can write to the block
-    CommonUtils.setLocalFileStickyBit(mLocalFilePath);
-    LOG.info(mLocalFilePath + " was created!");
-
     mBuffer = ByteBuffer.allocate(mUserConf.FILE_BUFFER_BYTES + 4);
   }
 
   private synchronized void appendCurrentBuffer(byte[] buf, int offset, int length)
       throws IOException {
-    if (!mTachyonFS.requestSpace(length)) {
+    boolean reqResult = false;
+    if (mLocationInfo == null) {
+      mLocationInfo = mTachyonFS.requestSpace(length);
+      reqResult = (mLocationInfo != null);
+    } else {
+      reqResult = mTachyonFS.requestSpace(mLocationInfo.getStorageDirId(), length);
+    }
+    if (!reqResult) {
       mCanWrite = false;
 
       String msg =
@@ -110,6 +103,17 @@ public class BlockOutStream extends OutStream {
               + mFile.mFileId + " " + mBlockId;
 
       throw new IOException(msg);
+    }
+    if (mLocalFilePath == null) {
+      String localTempFolder = createUserLocalTempFolder();
+      mLocalFilePath = CommonUtils.concat(localTempFolder, mBlockId);
+      mLocalFile = new RandomAccessFile(mLocalFilePath, "rw");
+      mLocalFileChannel = mLocalFile.getChannel();
+      // change the permission of the temporary file in order that the worker can move it.
+      CommonUtils.changeLocalFileToFullPermission(mLocalFilePath);
+      // use the sticky bit, only the client and the worker can write to the block
+      CommonUtils.setLocalFileStickyBit(mLocalFilePath);
+      LOG.info(mLocalFilePath + " was created!");
     }
 
     MappedByteBuffer out = mLocalFileChannel.map(MapMode.READ_WRITE, mInFileBytes, length);
@@ -143,14 +147,37 @@ public class BlockOutStream extends OutStream {
       }
 
       if (mCancel) {
-        mTachyonFS.releaseSpace(mWrittenBytes - mBuffer.position());
-        new File(mLocalFilePath).delete();
-        LOG.info("Canceled output of block " + mBlockId + ", deleted local file " + mLocalFilePath);
-      } else {
-        mTachyonFS.cacheBlock(mBlockId);
+        if (mLocationInfo != null) { // if file was written
+          mTachyonFS.releaseSpace(mLocationInfo.getStorageDirId(),
+              mWrittenBytes - mBuffer.position());
+          new File(mLocalFilePath).delete();
+          LOG.info("Canceled output of block " + mBlockId + ", deleted local file "
+              + mLocalFilePath);
+        }
+      } else if (mLocationInfo != null) {
+        mTachyonFS.cacheBlock(mLocationInfo.getStorageDirId(), mBlockId);
       }
     }
     mClosed = true;
+  }
+
+  /**
+   * Create temporary folder for the user in some StorageDir
+   * 
+   * @return the path of the folder
+   * @throws IOException
+   */
+  private String createUserLocalTempFolder() throws IOException {
+    String localTempFolder = null;
+    if (mLocationInfo != null) {
+      localTempFolder = mTachyonFS.createAndGetUserLocalTempFolder(mLocationInfo.getPath());
+    }
+
+    if (localTempFolder == null) {
+      mCanWrite = false;
+      throw new IOException("Failed to create temp user folder for tachyon client.");
+    }
+    return localTempFolder;
   }
 
   @Override
